Extract a conventions trait.
Having a trait with some commonly used inflections/conventions makes the
Shell class a bit cleaner and contain less junk. Having a trait also
allows userland code and other parts of the framework to re-use this
piece.
